Several fixes for lbaas2 (#712)
* [lbaas2] enable deletion of whole loadbalancers including their children objects

* [lbaas2] use correct policy <-> protocol mapping

* [lbaas2] don't allow insert_headers for https protocol (treat like tcp)

+++ plugins/lbaas2/app/javascript/lib/hooks/useLoadbalancer.js
@@ -87,7 +87,7 @@ const useLoadbalancer = () => {
   
   const deleteLoadbalancer = (name, id) => {
     return new Promise((handleSuccess,handleErrors) => {
-      confirm(<React.Fragment><p>Do you really want to delete following Load Balancer?</p><p>{createNameTag(name)} <b>id:</b> {id}</p></React.Fragment>).then(() => {
+      confirm(<React.Fragment><p>Do you really want to delete following Load Balancer and all related objects?</p><p>{createNameTag(name)} <b>id:</b> {id}</p></React.Fragment>).then(() => {
         return ajaxHelper.delete(`/loadbalancers/${id}`)
         .then( (response) => {
           dispatch({type: 'REQUEST_REMOVE_LOADBALANCER', id: id})
